refactor: clean up routes and navigation
